EuiFlyout[X] TS
