+++ src/components/flyout/flyout_body.test.tsx
@@ -1,16 +1,13 @@
 import React from 'react';
 import { render } from 'enzyme';
-import { requiredProps } from '../../test';
+import { requiredProps } from '../../test/required_props';
 
 import { EuiFlyoutBody } from './flyout_body';
 
 describe('EuiFlyoutBody', () => {
   test('is rendered', () => {
-    const component = render(
-      <EuiFlyoutBody {...requiredProps} />
-    );
+    const component = render(<EuiFlyoutBody {...requiredProps} />);
 
-    expect(component)
-      .toMatchSnapshot();
+    expect(component).toMatchSnapshot();
   });
 });
