+++ src/components/flyout/flyout_footer.test.tsx
@@ -1,16 +1,13 @@
 import React from 'react';
 import { render } from 'enzyme';
-import { requiredProps } from '../../test';
+import { requiredProps } from '../../test/required_props';
 
 import { EuiFlyoutFooter } from './flyout_footer';
 
 describe('EuiFlyoutFooter', () => {
   test('is rendered', () => {
-    const component = render(
-      <EuiFlyoutFooter {...requiredProps} />
-    );
+    const component = render(<EuiFlyoutFooter {...requiredProps} />);
 
-    expect(component)
-      .toMatchSnapshot();
+    expect(component).toMatchSnapshot();
   });
 });
